[picasaweb] Fixed declaration of gdata_picasaweb_file_get_size

+++ gdata/services/picasaweb/gdata-picasaweb-file.h
@@ -74,7 +74,7 @@ const gchar *gdata_picasaweb_file_get_album_id (GDataPicasaWebFile *self);
 void gdata_picasaweb_file_set_album_id (GDataPicasaWebFile *self, const gchar *album_id); /* TODO should we have a more obvious _move() API too? nah */
 guint gdata_picasaweb_file_get_width (GDataPicasaWebFile *self);
 guint gdata_picasaweb_file_get_height (GDataPicasaWebFile *self);
-gulong gdata_picasaweb_file_get_size (GDataPicasaWebFile *self);
+gsize gdata_picasaweb_file_get_size (GDataPicasaWebFile *self);
 const gchar *gdata_picasaweb_file_get_client (GDataPicasaWebFile *self);
 void gdata_picasaweb_file_set_client (GDataPicasaWebFile *self, const gchar *client);
 const gchar *gdata_picasaweb_file_get_checksum (GDataPicasaWebFile *self);
